% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{print.palette}
\alias{print.palette}
\title{Print palette}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{Palette}

\item{...}{}
}
\value{
Image?
}
\description{
Print palette
}
\examples{

oe_palette("blue_grad")
}
